Move convertCommandAndArgsToShell into Utilities.
